backend/Analyse/requirements.txt
# Dépendances principales
pandas==2.0.3
pdfplumber==0.10.3
python-dotenv==1.0.0

# Google Gemini API - Version compatible avec votre script
google-genai==1.63.0

# Flask pour le serveur API
Flask==2.3.3
Flask-CORS==4.0.0
werkzeug==2.3.7

# Utilitaires
requests==2.31.0